Refactor: Split Github into files
