+++ core/src/com/unciv/ui/screens/modmanager/ModInfoAndActionPane.kt
@@ -4,6 +4,7 @@ import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.scenes.scene2d.ui.Image
 import com.badlogic.gdx.scenes.scene2d.ui.Table
+import com.unciv.logic.github.GitHubData
 import com.unciv.models.metadata.BaseRuleset
 import com.unciv.models.ruleset.Ruleset
 import com.unciv.models.ruleset.validation.ModCompatibility
@@ -34,7 +35,7 @@ internal class ModInfoAndActionPane : Table() {
     /** Recreate the information part of the right-hand column
      * @param repo: the repository instance as received from the GitHub api
      */
-    fun update(repo: Github.Repo) {
+    fun update(repo: GitHubData.Repo) {
         isBuiltin = false
         enableVisualCheckBox = false
         update(
