+++ core/src/com/unciv/ui/screens/modmanager/ModUIData.kt
@@ -1,9 +1,9 @@
 package com.unciv.ui.screens.modmanager
 
+import com.unciv.logic.github.GitHubData
 import com.unciv.models.metadata.ModCategories
 import com.unciv.models.ruleset.Ruleset
 import com.unciv.models.translations.tr
-import com.unciv.logic.github.Github
 
 /** Helper class holds combined mod info for ModManagementScreen, used for both installed and online lists.
  *
@@ -15,7 +15,7 @@ internal class ModUIData private constructor(
     val name: String,
     val description: String,
     val ruleset: Ruleset?,
-    val repo: Github.Repo?,
+    val repo: GitHubData.Repo?,
     var isVisual: Boolean = false,
     var hasUpdate: Boolean = false
 ) {
@@ -27,7 +27,7 @@ internal class ModUIData private constructor(
         ruleset, null, isVisual = isVisual
     )
 
-    constructor(repo: Github.Repo, isUpdated: Boolean): this (
+    constructor(repo: GitHubData.Repo, isUpdated: Boolean): this (
         repo.name,
         (repo.description ?: "-{No description provided}-".tr()) +
                 "\n" + "[${repo.stargazers_count}]✯".tr(),
